rossmann-telegram-api/requirements.txt
Flask==2.1.0
inflection==0.5.1
numpy==1.22.3
pandas==1.4.1
requests==2.22.0
Pillow==9.0.1
seaborn==0.11.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
python-telegram-bot==13.11